+++ resources/views/livewire/messages/search.blade.php
@@ -15,7 +15,7 @@
             <ul class="bg-white dark:bg-gray-900 divide-y dark:divide-white/10">
                 @foreach($messages as $message)
                     <li wire:key="{{ $message->id }}" class="hover:bg-gray-100 hover:bg-gray-100 dark:hover:bg-white/10">
-                        <a wire:navigate href="{{ Messages::getUrl(tenant: filament()->getTenant()) . '/' . $message->inbox->id }}">
+                        <a wire:navigate href="{{ Messages::getPageUrl(tenant: filament()->getTenant()) . '/' . $message->inbox->id }}">
                             <div class="grid grid-cols-[--cols-default] lg:grid-cols-[--cols-lg] p-3" style="--cols-default: repeat(1, minmax(0, 1fr)); --cols-lg: repeat(5, minmax(0, 1fr));">
                                 <div style="--col-span-default: span 4 / span 4;" class="col-[--col-span-default]">
                                     <div class="flex gap-3">